+++ src/main/java/insane96mcp/enhancedai/modules/Modules.java
@@ -14,6 +14,7 @@ public class Modules {
 	public static Module drowned;
 	public static Module ghast;
 	public static Module shulker;
+	public static Module slime;
 	public static Module golem;
 	public static Module pillager;
 	public static Module enderman;
@@ -31,6 +32,7 @@ public class Modules {
 		drowned = Module.Builder.create(Ids.DROWNED, "Drowneds", ModConfig.Type.COMMON, Config.builder).build();
 		ghast = Module.Builder.create(Ids.GHAST, "Ghasts", ModConfig.Type.COMMON, Config.builder).build();
 		shulker = Module.Builder.create(Ids.SHULKER, "Shulkers", ModConfig.Type.COMMON, Config.builder).build();
+		slime = Module.Builder.create(Ids.SLIME, "Slimes", ModConfig.Type.COMMON, Config.builder).build();
 		pillager = Module.Builder.create(Ids.ILLAGER, "Illagers", ModConfig.Type.COMMON, Config.builder).build();
 		golem = Module.Builder.create(Ids.PETS, "Pets", ModConfig.Type.COMMON, Config.builder).build();
 		enderman = Module.Builder.create(Ids.ENDERMAN, "Endermen", ModConfig.Type.COMMON, Config.builder).build();
@@ -49,6 +51,7 @@ public class Modules {
 		public static final String DROWNED = EnhancedAI.RESOURCE_PREFIX + "drowned";
 		public static final String GHAST = EnhancedAI.RESOURCE_PREFIX + "ghast";
 		public static final String SHULKER = EnhancedAI.RESOURCE_PREFIX + "shulker";
+		public static final String SLIME = EnhancedAI.RESOURCE_PREFIX + "slime";
 		public static final String ILLAGER = EnhancedAI.RESOURCE_PREFIX + "illager";
 		public static final String PETS = EnhancedAI.RESOURCE_PREFIX + "pets";
 		public static final String ENDERMAN = EnhancedAI.RESOURCE_PREFIX + "enderman";